Split portfolio hook
